@@ -4,9 +4,11 @@
   "private": true,
   "homepage": "./",
   "dependencies": {
+    "@mui/icons-material": "^5.0.1",
+    "@mui/types": "^7.0.0",
     "@naisutech/react-tree": "^2.0.11",
-    "@rjsf/core": "^2.5.1",
-    "@rjsf/material-ui": "^2.5.1",
+    "@rjsf/core": "^3.1.0",
+    "@rjsf/material-ui": "^3.1.0",
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
@@ -21,19 +23,18 @@
     "d3": "^6.7.0",
     "d3-celestial": "^0.7.35",
     "d3-force": "^2.1.1",
-    "material-table": "^1.69.3",
     "npm-force-resolutions": "^0.0.10",
     "query-string": "^7.0.0",
-    "react": "^16.8.0",
+    "react": "^17.0.2",
     "react-beautiful-dnd": "^13.0.0",
     "react-dnd": "^14.0.3",
     "react-dnd-html5-backend": "^14.0.1",
-    "react-dom": "^16.8.0",
+    "react-dom": "^17.0.2",
     "react-force-graph": "^1.41.5",
     "react-json-view": "^1.21.3",
     "react-jsonschema-form": "^1.8.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^4.0.3",
     "react-sizeme": "^3.0.1",
     "react-spring": "^9.2.4",
     "react-uuid": "^1.0.2",
@@ -71,9 +72,9 @@
     "bezier-js": "4.0.3"
   },
   "devDependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/styles": "^4.11.3",
+    "@mui/icons-material-material": "^5.0.1",
+    "@mui/material": "^5.0.1",
+    "@mui/styles": "^5.0.1",
     "@types/d3": "^6.7.0",
     "@types/react-jsonschema-form": "^1.7.4",
     "@types/react-router-dom": "^5.1.7",
